--- Auto-Git Commit ---

+++ database/migrations/2019_08_29_161256_create_subscriptions_table.php
@@ -17,8 +17,8 @@ class CreateSubscriptionsTable extends Migration
             $table->bigIncrements('id');
             $table->string('date')->nullable($value = true);
             $table->string('image')->nullable($value = true);
-            $table->unsignedInteger('fannie_id')->nullable($value = true);
-            $table->unsignedBigInteger('subscription_id')->nullable($value = true);
+            $table->unsignedBigInteger('fannie_id')->nullable($value = true);
+             $table->unsignedBigInteger('subscription_id')->nullable($value = true);
             $table->foreign('fannie_id')->references('id')->on('users')->onDelete('set null'); 
             $table->foreign('subscription_id')->references('id')->on('subscription_types')->onDelete('set null'); 
             $table->timestamps();
